requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-filter==24.2
django-import-export==4.3.14
django-simple-history==3.6.0
python-decouple==3.8
Pillow==12.1.0
ruff==0.8.0